change balance's type to Double

+++ src/main/java/thirdSet/serialization/models/Client.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class Client implements Serializable {
     private final String username;
-    private final Integer balance;
+    private final Double balance;
     private final Integer age;
     private static final long serialVersionUID = -8237181389219902238L;
 
-    public Client(String username, Integer balance, Integer age) {
+    public Client(String username, Double balance, Integer age) {
         this.username = username;
         this.balance = balance;
         this.age = age;
